fix lapor hr

+++ app/Http/Controllers/OptionsController.php
@@ -9,6 +9,7 @@ use App\Models\Department;
 use Illuminate\Http\Request;
 use App\Models\EmployeeStatus;
 use App\Models\Holiday;
+use App\Models\LaporHr;
 use App\Models\LaporHrCategory;
 use App\Models\OfficeLocation;
 
@@ -367,6 +368,14 @@ class OptionsController extends Controller
 
     public function laporHrCategoryDelete ($id) {
         $laporHrCategory = LaporHrCategory::find($id);
+        $exist = LaporHr::where('category_id', $laporHrCategory);
+        if($exist) {
+            return response()->json([
+                'success' => false,
+                'message' => 'Data yang sudah digunakan, tidak bisa dihapus!',
+                'redirect' => route('options.list')
+            ]);
+        }
         $laporHrCategory->delete();
         return response()->json([
             'success' => true,
